Make plugins inherit from Plugin interface and plugin factories from PluginFactory interface
This will allow us to add common attributes to the plugin interface,
such as the plugin name, version, etc.

The factories now return a raw pointer because smart pointers don't meet
the covariance requirements for return type overriding. We should store
the raw pointer plugin we get from the factory in a smart pointer.