Add core and separate responsibilities
This commit makes MainWindow use multiple inheritance on its UI component.
This allows for accessing UI elements directly in the class which prevents something like a PluginManager to have to fetch the MainWindow from Core, and then fetch the UI from MainWindow.

The main window UI is no longer passed to LoadPlugins() as this function should not need to depend on a UI.

The pluginTriggered slot no longer needs to figure out what kind of plugin factory has been triggered. It just produces the plugin that belongs to the factory.

The main menu is now in charge of adding elements to its user interface. Adding UI elements to the window from any other class is not allowed.