Fixes some mean shift issues
Fixes some issues in mean shift clustering, particularly in large and unevenly distributed data, that occured due to changes from original implementation.
Adds a postprocessing step that merges clusters where the center (detected density peak) is not included in the cluster, with the corresponding cluster containing the center (This can only happen when the mean shift has not nicely converged to a single peak).