Replace private `PointData::VectorHolder` class with `std::variant`
`PointData::VectorHolder` "mimicked" a C++17 `std::variant` of `std::vector` types. Now that the ManiVault core is upgraded to C++17, it is preferred to use `std::variant` instead, to ease further maintenance.

Note that this commit does not change the interface (the API) of `PointData`. It is basically just a code clean-up!