Fix Font Awesome icons on Mac OS and Linux (#654)
* Add some debug printing

* Try only loading solid

* Only load Font Awesome regular fonts on Windows

The Mac OS and Linux font family names for regular and solid are the same: Font Awesome X Free. This makes it impossible to select the correct font family and assign the proper icon. At this point in time, we only use the solid icons, so we will only load the solid icons on Mac OS and Linux.

* Remove debug printing to console