Fix shutdown error by making SecureNetworkAccessManager a leaky singleton
When the application is shut down, a non-Qt thread will cleanup the living objects.
Since SecureNetworkAccessManager is not parented to any Qt object, Qt will not try to clean it up.
Hence it is being cleaned from a different thread than the Qt mainthread, which will cause an error.
Parenting it to a Qt object allows Qt to clean it up, but requires that we meticulously clean it up before it is deleted.
Most stable solution at the moment is to let the memory leak, and let the OS reclaim the memory on process exit.