Decomission custom splash screen for published projects (#1168)
* Comment out splash screen action in publishProject

The call to add the splash screen action to settingsGroupAction in publishProject was commented out to prevent it from being included during project publishing.

* Update splash screen action comment

Clarified that the splash screen action will be removed in the future due to application-wide splash screen customization.

* Deprecate project-level splash screen action

Removed accessors for SplashScreenAction from Project, and marked SplashScreenAction in ProjectMetaAction as deprecated in favor of application-wide splash screen customization.

* Remove splash screen action from publishProject

Eliminated handling of splashScreenEnabledAction in ProjectManager::publishProject, simplifying the project publishing logic to only manage the readOnlyAction.